% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read.gaf}
\alias{read.gaf}
\title{GAF Reader Function}
\usage{
read.gaf(gaf_file)
}
\arguments{
\item{gaf}{Path to a GAF (or compressed gaf) file}
}
\value{
A data frame containing the information parsed from the GAF file.
}
\description{
Reads a GAF format file and returns a data table.
}
\examples{
pig.gaf <- read.gaf("goa_pig_gaf.gz")
head(pig.gaf)
}
