% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read.gpa}
\alias{read.gpa}
\title{GPA Reader Function}
\usage{
read.gpa(gpa_file)
}
\arguments{
\item{gaf}{Path to a GPA (or compressed GPA) file}
}
\value{
A data frame containing the information parsed from the GPA file.
}
\description{
Reads a GPA format file and returns a data table.
}
\examples{
pig.gpa <- read.gpa("goa_pig_gpa.gz")
head(pig.gpa)
}
