% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read.gpi}
\alias{read.gpi}
\title{GPI Reader Function}
\usage{
read.gpi(gpi_file)
}
\arguments{
\item{gaf}{Path to a GPI (or compressed GPI) file}
}
\value{
A data frame containing the information parsed from the GPI file.
}
\description{
Reads a GPI format file and returns a data table.
}
\examples{
pig.gpi <- read.gpi("goa_pig_gpi.gz")
head(pig.gpi)
}
